+++ src/IDs/PanCard/PanCard.css
