+++ src/IDs/PanCard/assets/national-emblem.png
